server/requirements.txt
Flask==3.0.*
Flask-SocketIO==5.*
python-dotenv